+++ src/cite_runner/config.py
@@ -12,20 +12,20 @@ from rich.logging import RichHandler
 from . import models
 
 
-class TeamEngineRunnerSettings(BaseSettings):
+class CiteRunnerSettings(BaseSettings):
     model_config = SettingsConfigDict(
-        env_prefix="TEAMENGINE_RUNNER__",
+        env_prefix="CITE_RUNNER__",
         env_nested_delimiter="__"
     )
-    default_json_serializer: str = "ogc_cite_action.serializers.simple.to_json"
-    default_markdown_serializer: str = "ogc_cite_action.serializers.simple.to_markdown"
-    default_parser: str = "ogc_cite_action.parsers.earl.parse_test_suite_result"
+    default_json_serializer: str = "cite_runner.serializers.simple.to_json"
+    default_markdown_serializer: str = "cite_runner.serializers.simple.to_markdown"
+    default_parser: str = "cite_runner.parsers.earl.parse_test_suite_result"
     extra_templates_path: str | None = None
 
     # ogcapi_features_1_0_parser: str = (
-    #     "ogc_cite_action.teamengine_runner.parse_test_suite_result")
+    #     "cite_runner.teamengine_runner.parse_test_suite_result")
     # ogcapi_features_1_0_markdown_serializer: str = (
-    #     "ogc_cite_action.teamengine_runner.serialize_test_suite_result")
+    #     "cite_runner.teamengine_runner.serialize_test_suite_result")
     simple_serializer_template: str = "test-suite-result.md"
 
 
@@ -37,11 +37,11 @@ class CliContext(pydantic.BaseModel):
     debug: bool = False
     jinja_environment: jinja2.Environment = jinja2.Environment()
     network_timeout_seconds: int = 20
-    settings: TeamEngineRunnerSettings
+    settings: CiteRunnerSettings
 
 
-def get_settings() -> TeamEngineRunnerSettings:
-    return TeamEngineRunnerSettings()
+def get_settings() -> CiteRunnerSettings:
+    return CiteRunnerSettings()
 
 
 def get_context(
@@ -57,9 +57,9 @@ def get_context(
     )
 
 
-def _get_jinja_environment(settings: TeamEngineRunnerSettings) -> jinja2.Environment:
+def _get_jinja_environment(settings: CiteRunnerSettings) -> jinja2.Environment:
     loaders = [
-        jinja2.PackageLoader("ogc_cite_action", "templates"),
+        jinja2.PackageLoader("cite_runner", "templates"),
     ]
     if settings.extra_templates_path is not None:
         loaders.append(jinja2.FileSystemLoader(settings.extra_templates_path))
