+++ src/cite_runner/exceptions.py
@@ -1,0 +1,2 @@
+class CiteRunnerException(BaseException):
+    ...
