--- src/ogc_cite_action/exceptions.py
@@ -1,2 +1,0 @@
-class OgcCiteActionException(BaseException):
-    ...
