+++ tests/test_parsers_earl.py
@@ -1,4 +1,4 @@
-from ogc_cite_action.parsers import earl
+from cite_runner.parsers import earl
 
 
 def test_parse_test_suite_result_ogcapi_features_1_0(
